tools: scheduler: refactor as a lib instead of main package
