+++ toolkit/tools/depsearch/depsearch.go
@@ -9,14 +9,14 @@ import (
 	"path/filepath"
 	"sort"
 
+	"gonum.org/v1/gonum/graph"
+	"gopkg.in/alecthomas/kingpin.v2"
+
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/exe"
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/file"
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/logger"
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/pkggraph"
-	"github.com/microsoft/CBL-Mariner/toolkit/tools/scheduler/schedulerutils"
-
-	"gonum.org/v1/gonum/graph"
-	"gopkg.in/alecthomas/kingpin.v2"
+	"github.com/microsoft/CBL-Mariner/toolkit/tools/pkg/scheduler/schedulerutils"
 )
 
 const (
