+++ toolkit/tools/pkg/graph/pkgfetcher/pkgfetcher.go
@@ -5,15 +5,15 @@ import (
 	"path/filepath"
 	"strings"
 
+	"gonum.org/v1/gonum/graph"
+
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/logger"
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/packagerepo/repocloner/rpmrepocloner"
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/packagerepo/repoutils"
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/pkggraph"
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/pkgjson"
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/rpm"
-	"github.com/microsoft/CBL-Mariner/toolkit/tools/scheduler/schedulerutils"
-
-	"gonum.org/v1/gonum/graph"
+	"github.com/microsoft/CBL-Mariner/toolkit/tools/pkg/scheduler/schedulerutils"
 )
 
 func (cfg *Config) ResolvePackages() error {
