+++ toolkit/tools/pkg/scheduler/schedulerutils/buildworker.go
@@ -9,13 +9,14 @@ import (
 	"sync"
 	"time"
 
+	"gonum.org/v1/gonum/graph"
+	"gonum.org/v1/gonum/graph/traverse"
+
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/logger"
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/pkggraph"
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/retry"
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/sliceutils"
-	"github.com/microsoft/CBL-Mariner/toolkit/tools/scheduler/buildagents"
-	"gonum.org/v1/gonum/graph"
-	"gonum.org/v1/gonum/graph/traverse"
+	"github.com/microsoft/CBL-Mariner/toolkit/tools/pkg/scheduler/buildagents"
 )
 
 // BuildChannels represents the communicate channels used by a build agent.
